Added support to determine requested parameters
There are two sets of settigns that have to be recorded at runtime:
The settings contained in a file and the settings requested by a plugin.
If, for example, a plugin is delivered in binary form, the runtime has
to have a way to determine and communicate parameters needed by a
plugin.